Remove object.ListView
Move methods to view.ListView type.
